fix todo: merge pkg/api/v1/node/util.go with pkg/util/node/node.go
